Do not subtract an offset by default
This changes how the global offset subtraction is done in
ctapipe_io_lst.
Instead of subtracting `offset` from both the waveform and the drs4
pedestal, we now rely on the offset being part of the drs4 pedestal.

This means, that when the drs4 pedestal calibration is used, no offset
needs to be given, making it possible to analyze runs with different
offsets without adapting the configuration.